6.10-r9492_tbs_and_lots_of_stuff

+++ GDO/User/GDO_User.php
@@ -190,7 +190,7 @@ final class GDO_User extends GDO
 	{
 	    $level = $this->getVar('user_level');
 	    $permLevel = $this->getPermissionLevel();
-	    return max($level, $permLevel);
+	    return (int)max($level, $permLevel);
 	}
 	
 	public function getPermissionLevel()
